pyproject.toml
[project]
name = "mdexec"
version = "0.1.0"
description = "Lightweight Markdown executor for embedded code blocks"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "David Thigpen" }]
dependencies = [
    "markdown-it-py>=3.0.0",
    "pyyaml>=6.0",
]
[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pytest==8.3.5",
    "pytest-cov==6.1.1",
]

[project.scripts]
mdexec = "mdexec.main:main"

[tool.ruff]
# Same as Black
line-length = 88
indent-width = 4

target-version = "py39"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
